requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-WTF
email-validator
python-dotenv
openai
gunicorn
